(smach_states) migrate WM to folder
