+++ robot_smach_states/src/robot_smach_states/world_model/__init__.py
@@ -1,0 +1,2 @@
+from .volumes import *
+from .world_model import *
