+++ robot_smach_states/test/test_world_model.py
@@ -8,7 +8,7 @@ from ed.volume import BoxVolume
 
 from robot_skills import Mockbot
 from robot_skills.classification_result import ClassificationResult
-from robot_smach_states.world_model import SegmentObjects
+from robot_smach_states.world_model.world_model import SegmentObjects
 import robot_smach_states.util.designators as ds
